guard unsafe with build tag
