+++ internal/fastnetip/uint128_test.go
@@ -1,3 +1,5 @@
+//go:build unsafe
+
 package fastnetip
 
 import "testing"
